+++ CHANGELOG.md
@@ -15,6 +15,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Remove /demo folder, incorporate some examples into Advanced Configuration docs
 - Added instructions to run Widoco.
 - Tidied up some circular dependencies on command-line constants.
+- Updates to new namespace for the Artifact Generator's own vocab, and picks up
+  major YAML reorganization.
 - Removed type 'any' from generated vocabs in TS
 
 ## 0.13.4